Add support for queryRenderedFeatures in Qt interface
Implements: https://github.com/maplibre/maplibre-native-qt/issues/219

A common feature in interactive maps is the ability to click and
retrieve the feature located beneath the mouse cursor. This
functionality is already integrated in many other MapLibre
interfaces, and this commit attempts to bring similar support to
the Qt interface.

Due to the use of private structures, several forwarding methods
were necessary — something I'm not particularly fond of. A review
of this approach is recommended to see if a cleaner solution can
be found.

Additional include directives were added in CMakeLists.txt to
ensure proper MOC generation. It would be worth reviewing whether
these can be simplified.

An example has been added which allows users to click on the map
and retrieve feature properties. Currently, the mouse event
interface directly provides coordinates, which is not ideal for
this use case. We may need to revise the API to accommodate this
more appropriately.

Note that the features returned are of type mbgl::Feature. It’s
worth considering whether this is what the end-user expects.